+++ test/e2e-v2/script/prepare/setup-e2e-shell/install-swctl.sh
@@ -26,6 +26,5 @@ if ! command -v swctl &> /dev/null; then
   mkdir -p $BASE_DIR/swctl && cd $BASE_DIR/swctl
   curl -kLo skywalking-cli.tar.gz https://github.com/apache/skywalking-cli/archive/${SW_CTL_COMMIT}.tar.gz
   tar -zxf skywalking-cli.tar.gz --strip=1
-  utype=$(uname | awk '{print tolower($0)}')
-  make $utype-amd64 && mv bin/swctl-*-$utype-amd64 $BIN_DIR/swctl
+  make install DESTDIR=$BIN_DIR
 fi